backend/requirements.txt
# backend/requirements.txt

# Core Web Framework
Flask==2.2.5

# Database ORM
SQLAlchemy==1.4.49
psycopg2==2.9.7

# Payment Integration
stripe==5.8.0

# AWS SDK
boto3==1.28.57

# PDF Generation
reportlab==4.0.4
WeasyPrint==59.0

# LLM Integration
langchain==0.1.7
openai==1.3.5

# Testing and Development Tools
pytest==7.4.2
pytest-cov==4.1.0
flake8==6.1.0

# Utility Libraries
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0

# Security
cryptography==41.0.7

# Ensure Compatibility with Python 3.9+
python_version >= "3.9"
